refactor(web): extract ChatForm side-effect logic into hooks
Split four independent concerns out of ChatForm (653 → 453 LOC):

- useDraftPersistence(storageKey) — sessionStorage-backed input draft
- useSpeechRecognition({onTranscript}) — Web Speech API + mic state
- useNotionCredentials() — /api/credentials fetch + readiness flag
- useNotionSave({messages}) — per-message Notion append + state map

No behavior changes; all 85 tests still pass and the existing test
suite covers the new hooks transitively via ChatForm. This is step 1
of a 4-step refactor that prepares useChatStream extraction (step 3)
before Issue #112 (SSE resume) lands.
